fix: cross-table NFC UID dedup; fix: API float->Decimal for money amounts
- Added _ensure_unique_card_uid() — checks all 4 card tables
  before any card assign, preventing cross-table UID collisions (H1)
- parent_payments, parent_portal, platform_commission: removed float()
  casts on money values; chapa_service now accepts Decimal (C5 resolution)
- NFC cards: school_id column + migration + service updates
- License-server: auth on /verify, /activate, /school/{id}
- Bulk NFC assign: requires CARD_PRINT_ASSIGN permission (H2)
- MFA: FINANCE/SUPER_ADMIN users must enable MFA before login

+++ backend/app/services/chapa_service.py
@@ -3,6 +3,7 @@ import os
 import json
 import hashlib
 import hmac
+from decimal import Decimal
 from typing import Optional, Dict, Any
 from datetime import datetime, timezone
 import httpx
@@ -42,7 +43,7 @@ def _get_headers(secret_key: str) -> Dict[str, str]:
 
 
 def initialize_payment(
-    amount: float,
+    amount: Decimal,
     currency: str,
     email: str,
     first_name: str,
